+++ CHANGES.md
@@ -1,7 +1,13 @@
 # Changelog
 
+## 1.10.0 (4/9/24)
+* Added: New sitewide locations for global content areas.
+* Added: New "After Footer" location. Also adds notice about using After Footer location with form plugins.
+* Changed: Updated the updater.
+* Fixed: Conditions on blog page were not working in some configurations.
+
 ## 1.9.6 (1/19/24)
-* Fixed: Encoded special characters were displaying on the front end in some configurations.
+* Fixed: Encoded special characters were displayed on the front end in some configurations.
 
 ## 1.9.5 (1/18/24)
 * Fixed: Custom post type archive conditions not working in some configurations.